LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/month-expected.txt

Inputs Month


PASS month type support on input element
FAIL The value attribute, if specified and not empty, must have a value that is a valid month string assert_equals: expected "2011-01" but got "1999-01"
FAIL The min attribute, if specified, must have a value that is a valid month string. assert_equals: expected "" but got "invalid_min"
FAIL The max attribute, if specified, must have a value that is a valid month string assert_equals: expected "2099-01" but got "2011-12"
PASS User agents must not allow the user to set the value to a non-empty string that is not a valid month string.
PASS Month value can be empty string.
PASS When value attribute has two digits year value, the value,which is invalid, must return empty string.
PASS When value is set with invalid value, the value must return empty string.
PASS When value is given invalid value to non-empty valid string, the value must be same as before.
PASS When step attribute is given invalid value, it must ignore the invalid value and use defaul value instead.
PASS Month should be <= 13: If the value of the element is not a valid month string, then set it to the empty string instead.
PASS Month should be > 0: If the value of the element is not a valid month string, then set it to the empty string instead.>
